todo-backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for DoQuanta Todo App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "psycopg2-binary>=2.9.11",
    
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
    "openai-agents>=0.1.0",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "python-dateutil>=2.9.0",
    "parsedatetime>=2.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]